# docs/api/openapi/storefront.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/storefront/addresses": {
      "get": {
        "operationId": "StorefrontAddressController_findAll",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddressResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the signed-in shopper's saved addresses",
        "tags": [
          "Addresses"
        ]
      },
      "post": {
        "operationId": "StorefrontAddressController_create",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Save an address without going through a cart",
        "tags": [
          "Addresses"
        ]
      }
    },
    "/storefront/addresses/{addressId}": {
      "patch": {
        "operationId": "StorefrontAddressController_update",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Address not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update one of the shopper’s saved addresses",
        "tags": [
          "Addresses"
        ]
      },
      "delete": {
        "description": "Past orders keep their own address snapshot and are unaffected.",
        "operationId": "StorefrontAddressController_remove",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedAddressDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Address not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one of the shopper’s saved addresses",
        "tags": [
          "Addresses"
        ]
      }
    },
    "/storefront/cart/{cartId}/addresses": {
      "get": {
        "description": "Empty until the cart has contact details and a first address is saved.",
        "operationId": "GuestAddressController_findAll",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddressResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "List the shopper's saved addresses",
        "tags": [
          "Cart"
        ]
      },
      "post": {
        "operationId": "GuestAddressController_create",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cart has no contact name plus email or phone yet"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Storefront not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Save an address to the shopper address book",
        "tags": [
          "Cart"
        ]
      }
    },
    "/storefront/cart/{cartId}/addresses/{addressId}": {
      "patch": {
        "operationId": "GuestAddressController_update",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Address not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Update a saved address",
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "description": "Past orders keep their own address snapshot and are unaffected.",
        "operationId": "GuestAddressController_remove",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedAddressDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Address not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Delete a saved address",
        "tags": [
          "Cart"
        ]
      }
    },
    "/storefront/cart/{cartId}/addresses/{addressId}/assign": {
      "patch": {
        "description": "Setting a shipping address clears any delivery selection, since the chosen method may not serve the new destination.",
        "operationId": "GuestAddressController_assign",
        "parameters": [
          {
            "name": "X-Merchant-Slug",
            "in": "header",
            "description": "Slug of the storefront being browsed",
            "required": true,
            "schema": {
              "type": "string",
              "example": "acme-store"
            }
          },
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCartAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed X-Merchant-Slug header"
          },
          "401": {
            "description": "Missing or invalid cart token"
          },
          "404": {
            "description": "Address not found"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ],
        "summary": "Use this address as the cart shipping or billing address",
        "tags": [
          "Cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "Storefront API",
    "description": "Storefront browsing, carts, shopper addresses (guest via cart token, or signed-in via bearer token), delivery options and tracking, checkout sessions, payment initiation, and payment status polling",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "checkout-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Checkout-Token"
      },
      "cart-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cart-Token"
      }
    },
    "schemas": {
...
      "CartLineDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "The cart item id, used to update or remove the line"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "string",
            "example": "19.99"
          },
          "totalPrice": {
            "type": "string",
            "example": "39.98"
          },
          "note": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "productId",
          "sku",
          "name",
          "quantity",
          "unitPrice",
          "totalPrice"
        ]
      },
...
      "CartDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CONVERTED",
              "MERGED",
              "ABANDONED",
              "EXPIRED"
            ]
          },
          "sourceChannel": {
            "type": "string",
            "enum": [
              "POS",
              "WEBSITE",
              "FACEBOOK",
              "INSTAGRAM",
              "TIKTOK"
            ]
          },
          "customerId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "ownerId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "The signed-in shopper this cart belongs to, once claimed. Null for a guest cart."
          },
          "customerName": {
            "type": "object",
            "nullable": true
          },
          "customerEmail": {
            "type": "object",
            "nullable": true
          },
          "customerPhone": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "subtotalAmount": {
            "type": "string",
            "example": "39.98"
          },
          "shippingAmount": {
            "type": "string",
            "example": "2.50"
          },
          "totalAmount": {
            "type": "string",
            "example": "42.48"
          },
          "itemCount": {
            "type": "number",
            "example": 2
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartLineDto"
            }
          },
          "shippingAddress": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "billingAddress": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "delivery": {
            "nullable": true,
            "description": "The chosen delivery method, re-quoted at read time",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryOptionDto"
              }
            ]
          },
          "checkoutSessionId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "sourceChannel",
          "currency",
          "subtotalAmount",
          "shippingAmount",
          "totalAmount",
          "itemCount",
          "items",
          "expiresAt"
        ]
      },
...
      "AddressResponseDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "object",
            "nullable": true
          },
          "recipientName": {
            "type": "string"
          },
          "phone": {
            "type": "object",
            "nullable": true
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": "object",
            "nullable": true
          },
          "city": {
            "type": "object",
            "nullable": true
          },
          "province": {
            "type": "object",
            "nullable": true
          },
          "postalCode": {
            "type": "object",
            "nullable": true
          },
          "country": {
            "type": "string",
            "example": "KH"
          },
          "latitude": {
            "type": "object",
            "nullable": true,
            "example": "11.5564000"
          },
          "longitude": {
            "type": "object",
            "nullable": true,
            "example": "104.9282000"
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "The signed-in shopper who owns this address. Null for a walk-in customer, who has no account."
          },
          "createdById": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "The staff member who saved it on the customer behalf, if any"
          },
          "isDefaultShipping": {
            "type": "boolean"
          },
          "isDefaultBilling": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "recipientName",
          "line1",
          "country",
          "id",
          "customerId",
          "isDefaultShipping",
          "isDefaultBilling",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 60,
            "example": "Home"
          },
          "recipientName": {
            "type": "string",
            "maxLength": 120
          },
          "phone": {
            "type": "string",
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "line1": {
            "type": "string",
            "maxLength": 200,
            "example": "12 Street 240"
          },
          "line2": {
            "type": "string",
            "maxLength": 200
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "example": "Phnom Penh"
          },
          "province": {
            "type": "string",
            "maxLength": 100
          },
          "postalCode": {
            "type": "string",
            "maxLength": 20
          },
          "country": {
            "type": "string",
            "example": "KH",
            "description": "ISO 3166-1 alpha-2 code"
          },
          "latitude": {
            "type": "number",
            "example": 11.5564
          },
          "longitude": {
            "type": "number",
            "example": 104.9282
          },
          "note": {
            "type": "string",
            "maxLength": 500
          },
          "isDefaultShipping": {
            "type": "boolean",
            "description": "Use as the customer's default shipping address. Omitted on their first address means yes."
          },
          "isDefaultBilling": {
            "type": "boolean",
            "description": "Use as the customer's default billing address."
          }
        },
        "required": [
          "recipientName",
          "line1",
          "country"
        ]
      },
      "UpdateAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 60,
            "example": "Home"
          },
          "recipientName": {
            "type": "string",
            "maxLength": 120
          },
          "phone": {
            "type": "string",
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "line1": {
            "type": "string",
            "maxLength": 200,
            "example": "12 Street 240"
          },
          "line2": {
            "type": "string",
            "maxLength": 200
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "example": "Phnom Penh"
          },
          "province": {
            "type": "string",
            "maxLength": 100
          },
          "postalCode": {
            "type": "string",
            "maxLength": 20
          },
          "country": {
            "type": "string",
            "example": "KH",
            "description": "ISO 3166-1 alpha-2 code"
          },
          "latitude": {
            "type": "number",
            "example": 11.5564
          },
          "longitude": {
            "type": "number",
            "example": 104.9282
          },
          "note": {
            "type": "string",
            "maxLength": 500
          },
          "isDefaultShipping": {
            "type": "boolean",
            "description": "Use as the customer's default shipping address. Omitted on their first address means yes."
          },
          "isDefaultBilling": {
            "type": "boolean",
            "description": "Use as the customer's default billing address."
          }
        }
      },
      "DeletedAddressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
      "AssignCartAddressDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "SHIPPING",
              "BILLING"
            ],
            "description": "Which slot on the cart this address fills"
          }
        },
        "required": [
          "type"
        ]
      },
...
    }
  }
}